Extract AbstractMantaEncryptedObjectInputStreamTest from MantaEncryptedObjectInputStreamTest
